[Parser] complete parsing module
