+++ syntax/src/parser/parsing/mod.rs
@@ -4,7 +4,7 @@ use super::CommandOrEnd;
 
 mod command;
 mod expression;
-mod operator;
+mod utils;
 mod pattern;
 mod ty;
 
@@ -13,29 +13,12 @@ pub use chumsky::prelude::*;
 
 pub use command::*;
 pub use expression::*;
-pub use operator::*;
+pub use utils::*;
 pub use pattern::*;
 pub use ty::*;
 
 pub type ErrorChumsky = chumsky::error::Rich<'static, Token, Span>;
 pub type ExtraChumsky = chumsky::extra::Err<ErrorChumsky>;
-
-// ===========================================================================
-// End of input
-// ===========================================================================
-
-/// parse end of input
-pub fn end_of_input() -> impl Parser<'static, TokenStream, cst::EndOfFile, ExtraChumsky> {
-    use cst::file::EndOfFileT;
-    select! {meta @ Meta{ value: TokenKind::EndOfInput, ..} =>
-        meta.map(|_| EndOfFileT())
-    }
-    .labelled("")
-}
-
-// ===========================================================================
-// End of input
-// ===========================================================================
 
 /// parse with lexer tokens
 pub fn parser() -> impl Parser<'static, TokenStream, CommandOrEnd, ExtraChumsky> {
